pyproject.toml
[tool.poetry]
name = "kubeq"
version = "0.1.0"
description = "Query your cluster"
authors = ["GregRos <work.gregr@gmail.com>"]
readme = "README.md"
license = "MIT"
homepage = "gh:GregRos/kubeq"
repository = "gh:GregRos/kubeq"
keywords = ["kubernetes", "query", "cli", "kubectl", "k8s"]
packages = [{ include = "kubeq" }]
[tool.poetry.dependencies]
python = "3.13.*"
pygments = "^2.19.1"
termcolor = "^2.5.0"
kubernetes = "^32.0.0"
kr8s = "0.20.3"
typeguard = "2.13.*"
aioreactive = "^0.20.0"
prettyprinttree = "^2.0.1"
python-box = "^7.3.2"
pyyaml = "^6.0.2"
httpx = "^0.28.1"

[tool.poetry.scripts]
test = "test.main:start"
start = "kubeq.main:start"
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
